Extract shared device browser functionality into shared component
